+++ src/Models/TrackAction.php
@@ -78,15 +78,4 @@ class TrackAction extends Tracking
     {
         return parent::query();
     }
-
-//    /**
-//     * Retrieve the 'model_key' key attribute as an integer.
-//     *
-//     * @param $value
-//     * @return int
-//     */
-//    public function getModelKeyAttribute($value): int
-//    {
-//        return intval($value);
-//    }
 }