+++ tests/Feature/Factories/Interfaces/FactoryAttributesTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sfneal\Tracking\Tests\Feature\Factories\Interfaces;
+
+// todo: consider moving to sfneal/mock-models?
+interface FactoryAttributesTest
+{
+    /** @test */
+    public function attributes_are_correct_types();
+}