+++ tests/Feature/Factories/Interfaces/FactoryFillablesTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Sfneal\Tracking\Tests\Feature\Factories\Interfaces;
+
+// todo: consider moving to sfneal/mock-models?
+interface FactoryFillablesTest
+{
+    /** @test */
+    public function fillables_are_correct_types();
+}