+++ CHANGELOG.md
@@ -15,3 +15,10 @@ All notable changes to `tracking` will be documented in this file
 - fix use of `activeUserID()` helper function in `TrackActivityEvent`
 - add sfneal/mock-models to dev requirements
 - fix migration paths in `TrackingServiceProvider`
+
+
+## 0.2.0 - 2021-04-14
+- add use of `env()` helper function in config file
+- fix use of sfneal/string-helpers helper functions by adding import of `StringHelpers`
+- fix issue with `TrackTrafficObserver` not being attached to `TrackTraffic` model
+- fix issue with `TrackTraffic` model not passing value to attribute setter in `TrackTrafficObserver`